feat: update CV files with new content and formats
This commit updates the CV files in both English and Portuguese, providing more detailed information and improved formatting.

The changes include:

-   Replacing the old `CV_HuannAlmeida_EN.pdf` with a new version that includes a professional summary, technical projects, tech stack, and updated experience and education sections.
-   Adding a new `Huann Almeida PTBR.docx` file, providing a Portuguese version of the CV in Word format.
-   Deleting the old `CV_HuannAlmeida_PT.pdf` file.
-   Adding a new `CV_HuannAlmeida_PTBR.pdf` file, providing a Portuguese version of the CV in PDF format.
-   Adding a new `Huann Almeida EN.docx` file, providing an English version of the CV in Word format.

These updates provide a more comprehensive and up-to-date representation of my skills and experience. The inclusion of both PDF and Word formats, as well as versions in both English and Portuguese, increases accessibility for potential employers.